# cmake: pre-fill rest of detection values for Windows