--- a/CMake/Platforms/WindowsCache.cmake
+++ b/CMake/Platforms/WindowsCache.cmake
@@ -194,6 +194,8 @@ if(NOT UNIX)
     set(HAVE_O_NONBLOCK 0)
     set(HAVE_IN_ADDR_T 0)
     set(STDC_HEADERS 1)
+
+    set(HAVE_SIZEOF_SUSECONDS_T 0)
   else()
     message("This file should be included on Windows platform only")
   endif()