--- a/CMakeLists.txt
+++ b/CMakeLists.txt
@@ -1068,7 +1068,7 @@ if(CURL_CA_PATH_SET AND
 endif()
 
 # Check for header files
-if(NOT UNIX)
+if(WIN32)
   check_include_file_concat("winsock2.h"     HAVE_WINSOCK2_H)
   check_include_file_concat("ws2tcpip.h"     HAVE_WS2TCPIP_H)
   check_include_file_concat("windows.h"      HAVE_WINDOWS_H)