@vszakats 2023-10-06
Also tested an experimental configure-vs-cmake feature-detection comparison.
Besides differently detected dependencies and some usual suspects, they were
identical, except `#define CURL_CA_PATH "/etc/ssl/certs"` which was picked up
by CMake, but missed by autotools:
https://github.com/curl/curl/actions/runs/6431509560/job/17464559513?pr=12044

macOS configure-vs-cmake patch:
```diff
diff --git a/.github/workflows/configure-vs-cmake.yml b/.github/workflows/configure-vs-cmake.yml
index 3131bc128..0f20801a3 100644
--- a/.github/workflows/configure-vs-cmake.yml
+++ b/.github/workflows/configure-vs-cmake.yml
@@ -27,7 +27,7 @@ on:
 permissions: {}
 
 jobs:
-  check:
+  check-linux:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
@@ -39,7 +39,28 @@ jobs:
 
     - name: run cmake
       run: |
-         mkdir build && cd build && cmake ..
+         cmake -B build
+
+    - name: compare generated curl_config.h files
+      run: ./scripts/cmp-config.pl lib/curl_config.h build/lib/curl_config.h
+
+  check-macos:
+    runs-on: macos-latest
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: install packages
+      run: |
+         while [[ $? == 0 ]]; do for i in 1 2 3; do brew update && brew install libtool autoconf automake && break 2 || { echo Error: wait to try again; sleep 10; } done; false Too many retries; done
+
+    - name: run configure --with-openssl
+      run: |
+         autoreconf -fi
+         ./configure --with-openssl
+
+    - name: run cmake
+      run: |
+         cmake -B build
 
     - name: compare generated curl_config.h files
       run: ./scripts/cmp-config.pl lib/curl_config.h build/lib/curl_config.h
```